requirements.txt
boto3
python-dotenv
PyPDF2
rapidfuzz
googlemaps
folium
openpyxl
pandas
pytest
